services/consumer/requirements.txt

kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
